Split EndpointFactory to ControllerFactory and EndpointInvoker
* ControllerFactory takes care of instantiating the controller and passing all dependencies
* EndpointInvoker takes care of Controller methods invoking flow
